Update supplied files to match my system
The supplied image and css files have been prefixed with 'rover-' to avoid any
future clashes with other projects. All my projects are hosted in a common root
on the server, and name conlifts are possible, especially with css/main.css.

The html file is updated to reference the new names and trimmed of the outomatic
additions from the project repo creator script. Likewise, the project_page.html
file is now index.html, but will remain as l11-ex01.html on the webserver.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>